+++ CHANGELOG.md
@@ -4,14 +4,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
-## [0.1.12]
+## [0.1.13]
 
 - Add missing ut for services
-  
+
+## [0.1.12]
+
+- Hash sensitive values in logs
+
 ## [0.1.11]
 
 - Add interceptor for communication mode toggle checking
- 
+
 ## [0.1.10]
 
 - Add interceptor for handling unhandled errors